wingification/event-driven-image-resizer/phase1/python/requirements.txt
wingsdk
Pillow
boto3
requests
